refacto: rename lib to predictsignauxfaibles
